Moved Message to root namespace
