+++ tests/LanguageServerTest.php
@@ -5,7 +5,7 @@ namespace LanguageServer\Tests;
 
 use PHPUnit\Framework\TestCase;
 use LanguageServer\LanguageServer;
-use LanguageServer\Factory\Message;
+use LanguageServer\Message;
 use LanguageServer\Protocol\{
     ClientCapabilities,
     TextDocumentSyncKind,
